Added exercise 2.13, corrected 2.12c and organized repo

+++ ficha1/cap2/2.11c.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 
 int main() {
-    int lastDen = 1;              
+    double lastDen = 1;              
     long double res = 1;                    // Adding 1 (the first member of the series manually)
     int x, reps;                            // User input
